Store / Stacks – Add stacks to store in backend-defined order instead of using the stack's id as index.
This enables us to respect the order of stacks when rendering them.
It makes it a bit harder to access a specific stack, but `.find` should be fine for now.
In a function component, performance could be improved by using a `useMemo()` hook to store the stacks of the current board and the currently selected stack, so we don't run an expensive lookup during each render.
However, it looks like this is performing well for now

This change assumes that the first stack is always the one with `order === 0`.